update loadcell treadmill
