+++ tests/treadmill_lc/info_s01_d01.txt
@@ -1,0 +1,2 @@
+Suj,Dia,Tent,BORG,Vel,Peso
+S01,01,T01,3,15,61.6
