+++ vaila/help/index.html
@@ -468,9 +468,9 @@
                     <div class="tool-links"><a href="#">-</a></div>
                 </div>
                 <div class="tool-card">
-                    <div class="tool-title">B6_r7_c3 - Load Cells</div>
-                    <div class="tool-links"><a href="loadcell_treadmill.html">HTML</a> | <a
-                            href="loadcell_treadmill.md">MD</a></div>
+                    <div class="tool-title">B6_r7_c3 - Treadmill LC</div>
+                    <div class="tool-links"><a href="treadmill_lc.html">HTML</a> | <a
+                            href="treadmill_lc.md">MD</a></div>
                 </div>
                 <div class="tool-card">
                     <div class="tool-title">B6_r7_c4 - <i>vailá</i></div>
@@ -717,7 +717,7 @@
         </div>
 
         <div class="footer">
-            <p class="muted">Generated based on <i>vailá</i> v0.3.67 layout. Generated on 29/06/2026.</p>
+            <p class="muted">Generated based on <i>vailá</i> v0.3.68 layout. Generated on 30/06/2026.</p>
         </div>
 
         <script>
